refactor: Update system id allocation script to use jsonb data type
The code changes in this commit refactor the system id allocation script in the utility/deploy/20241023_ALLOCATE_SYSTEM_IDS.sql file. The changes involve updating the data type of the p_data parameter from record to jsonb. This allows for more flexibility in handling and storing data within the script.