feat: Increase size of descriptive field in dimensions (#323)
This commit adds a new SQL script to alter the type of the descriptive field in the dimensions table. The size of the field is increased to accommodate longer descriptions. This change is temporary and has been merged into the SEAD_DATA_BASE_MODEL change request.

Issue: https://github.com/humlab-sead/sead_change_control/issues/323